Update dependencies and refactor code base and many improvements

+++ app-data/src/main/kotlin/co/anitrend/support/crunchyroll/data/api/helper/SessionHelper.kt
@@ -30,7 +30,7 @@ internal object SessionHelper {
         val builder = StringBuilder(length)
         repeat(length) {
             val random = Random.nextInt(start, end)
-            builder.append(whiteList[random].toChar())
+            builder.append(whiteList[random].toInt().toChar())
         }
         return builder.toString()
     }
